requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
cloudinary>=1.36.0
requests>=2.31.0
pdfkit>=1.0.0
weasyprint>=60.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dotenv>=1.0.0
